chore [datadog] update deprecated directive `datadog.apm.enabled`
When applying the datadog's helm chart, we got the following deprecation message:

```

The option `datadog.apm.enabled` is deprecated, please use `datadog.apm.portEnabled` to enable TCP communication to the trace-agent.
The option `datadog.apm.socketEnabled` is enabled by default and can be used to rely on unix socket or name-pipe communication.
```

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>